refactor: created custom error class

+++ src/storage/group/index.ts
@@ -1,10 +1,16 @@
 import AsyncStorage from '@react-native-async-storage/async-storage'
 
 import { GROUP_COLLECTION } from '@storage/store.config'
+import { AppError } from '@utils/AppError'
 
 export async function groupCreate(newGroupName: string) {
   try {
     const storagedGroups = await groupsGetAll()
+
+    const groupAlreadyExists = storagedGroups.includes(newGroupName)
+
+    if(groupAlreadyExists)
+      throw new AppError('Já existe um grupo cadastrado com esse nome.')
 
     const storage: string = JSON.stringify([...storagedGroups, newGroupName])
     await AsyncStorage.setItem(GROUP_COLLECTION, storage)
